requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
uvloop==0.17.0
